+++ fabric/src/main/java/nl/enjarai/a_good_place/fabric/AGoodPlaceImpl.java
@@ -44,7 +44,7 @@ public class AGoodPlaceImpl implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         WorldRenderEvents.AFTER_ENTITIES.register((context) -> {
-            BlocksParticlesManager.renderParticles(context.matrixStack(), context.tickDelta());
+            BlocksParticlesManager.renderParticles(context.matrixStack(), context.tickCounter().getGameTimeDeltaPartialTick(false));
         });
         ClientLifecycleEvents.CLIENT_STARTED.register(AGoodPlace::onSetup);
         ClientTickEvents.END_WORLD_TICK.register(BlocksParticlesManager::tickParticles);