+++ changelog.md
@@ -1,1 +1,7 @@
-changed mixin priority for mod compat
+## 1.21.1-1.2.8
+- Updated to Minecraft 1.21.1
+- Migrated from Forge to NeoForge
+- Updated to Java 21
+
+## Previous
+- changed mixin priority for mod compat